Add reference- and in-article link checking to the wikilink checker
The wikilink checker is updated with changes
from walavouchey/osu-wiki-tools@76f2b53ad7569455a69eba665f461c372899ab3f
to walavouchey/osu-wiki-tools@cdbfa960d9ef301a1c8158bf28e20696fd66290f
to make some more progress towards #6233. The full commit history along
with tests can be viewed in that repo.

The two main features of reference- and in-article link checking require
additional parsing logic to handle ignoring comments and getting the
correct identifier from headings and custom identifier tags. For
in-article links, the articles they point to are parsed for their
identifiers in order to check if the #link-fragment is correct, and if
not, all possible identifiers are printed as suggestions.

By default, all links from outdated articles are ignored. The plan is to
eventually turn on link checking for all files in CI runs when the
remaining errors have been fixed, so this is to prevent contributors
from having to edit already outdated articles.

Currently on master this this runs for about 40s with `--outdated` and
about 32s without on my Windows machine. There are 152 remaining errors
without counting outdated articles and 253 errors in total.